comment is now on BurgerPage
