+++ IBBS/Pages/BurgerPage.cshtml
@@ -10,6 +10,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>@ViewData["Title"] </title>
     <link rel="stylesheet" href="~/css/burgerpage.css" />
+  
     <link rel="stylesheet" href="~/lib/font-awesome/css/all.css" />
 </head>
 <body>
@@ -62,6 +63,63 @@
         </div>
     </div>
 
-    
+    <br />
+    <br />
+
+    <hr />
+    <div class="comments">
+
+        <div class="row">
+            <div class="col-md-4">
+                <form method="post">
+                    <div asp-validation-summary="ModelOnly" class="text-danger"></div>
+                    <h>Give your feedback for this Burger</h>
+                    <br />
+
+
+                    <div class="theComment">
+                        <!--<label asp-for="Comments.Comment" class="control-label"></label>-->
+                        <input asp-for="Comments.Comment" class="form-control" />
+                        <span asp-validation-for="Comments.Comment" class="btnComment"></span>
+
+                    </div>
+                    <br />
+                    <div class="comment">
+                        <input type="submit" value="Create" class="btnComment" />
+                    </div>
+                </form>
+            </div>
+            <hr />
+            <hr />
+        </div>
+    </div>
+
+
+    <!--
+    @foreach (var item in Model.commentList)
+    {
+
+        <p>
+            @Html.DisplayFor(modelItem => item.Comment)
+        </p>
+        <hr />
+        <br />
+
+    }
+    <hr />
+     -->
+
+
+
+
+
+
+
+
 </body>
 </html>
+
+
+@section Scripts {
+    @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
+}
