--- IBBS/Pages/issueNr13.cshtml
@@ -1,61 +1,0 @@
-﻿@page
-@model IBBS.Pages.issueNr13Model
-
-@{
-    ViewData["Title"] = "issueNr13";
-    // this is a scaffolded creat-page!
-}
-
-
-
-<hr />
-<div class="comments">
-
-    <div class="row">
-        <div class="col-md-4">
-            <form method="post">
-                <div asp-validation-summary="ModelOnly" class="text-danger"></div>
-                <h>Give your feedback for this Burger</h>
-                <br />
-
-
-                <div class="theComment">
-                    <!--<label asp-for="Comments.Comment" class="control-label"></label>-->
-                    <input asp-for="Comments.Comment" class="form-control" />
-                    <span asp-validation-for="Comments.Comment" class="btnComment"></span>
-
-                </div>
-                <br />
-                <div class="comment">
-                    <input type="submit" value="Create" class="btnComment" />
-                </div>
-            </form>
-        </div>
-        <hr />
-        <hr />
-    </div>
-</div>
-
-
-<!--
-@foreach (var item in Model.commentList)
-{
-
-    <p>
-        @Html.DisplayFor(modelItem => item.Comment)
-    </p>
-    <hr />
-    <br />
-
-}
-<hr />
- -->
-
-
-
-
-
-
-@section Scripts {
-    @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
-}
